+++ Nos/Views/NoteComposer/ImagePickerButton.swift
